refactor: runtime side effects
